review

+++ core/src/typedefs/fee_cents.rs
@@ -20,7 +20,7 @@ impl FeeCents {
 
     #[inline]
     pub const fn apply(&self, amt: u64) -> Option<AftFee> {
-        type F = sanctum_fee_ratio::Fee<Floor<Ratio<u32, u64>>>;
+        type F = sanctum_fee_ratio::Fee<Floor<Ratio<u32, u32>>>;
 
         let f = match F::new(Ratio {
             n: self.bp_cents,
